LayoutTests/inspector/console/console-css-warnings-expected.txt

Tests CSS warnings are properly exposed.

Unexpected CSS token: # console-css-warnings.html:10
Unexpected CSS token: { console-css-warnings.html:15
Unexpected CSS token: # console-css-warnings.html:18
Unexpected CSS token: # console-css-warnings.html:22
Unexpected CSS token: # console-css-warnings.html:32
Unexpected CSS token: { console-css-warnings.html:37
Unexpected CSS token: { console-css-warnings.html:38
Unexpected CSS token: { console-css-warnings.html:39
Unexpected CSS token: { console-css-warnings.html:40
Unexpected CSS token: * console-css-warnings.html:44
